requirements.txt
Flask 
Flask-SQLAlchemy 
Flask-Migrate 
Flask-Bcrypt 
PyJWT
psycopg[binary]
pydantic
python-dotenv